refactoring the weather shopper payment page
